Reorganize Unit Tests - Add Unit Tests for RelayCommands
