+++ MvvmControls/Mvvm.csproj
@@ -1,0 +1,109 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <TargetFrameworks>net472;net48;netcoreapp3.1;net5.0-windows</TargetFrameworks>
+        <RootNamespace>RFBCodeWorks</RootNamespace>
+        <AssemblyName>RFBCodeWorks.Mvvm.Controls</AssemblyName>
+        <LangVersion>9.0</LangVersion>
+        <Platforms>AnyCPU</Platforms>
+        <UseWPF>true</UseWPF>
+
+        <!-- Nuget Stuff -->
+        <GeneratePackageOnBuild Condition="$(Configuration)=='Release'">true</GeneratePackageOnBuild>
+        <PackageId>$(AssemblyName)</PackageId>
+        <Product>$(AssemblyName)</Product>
+        <PackageOutputPath>..\_OutputPackages\</PackageOutputPath>
+        <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
+        <PackageLicenseExpression></PackageLicenseExpression>
+        <DocumentationFile>bin\$(Configuration)\\$(AssemblyName).xml</DocumentationFile>
+        <Authors>Robert Brenckman</Authors>
+        <Company />
+        <Copyright>Copyright 2023</Copyright>
+        <Version>1.0.0.0</Version>
+        <PackageProjectUrl>https://github.com/RFBCodeWorks/MvvmControls</PackageProjectUrl>
+        <RepositoryUrl>https://github.com/RFBCodeWorks/MvvmControls</RepositoryUrl>
+        <PackageReadmeFile>Readme.md</PackageReadmeFile>
+        <PackageReleaseNotes>Fixed ICommand objects not reacting to CommandManager.RequerySuggested event</PackageReleaseNotes>
+
+        
+
+        <Description>
+            This is a Library designed to ease working within an MVVM framework, and is meant to consolidate boiler-plate ViewModel interactions for common WPF controls into pre-built classes.
+
+            ## Example - A simplified ComboBox Experience
+
+            A ComboBoxDefinition has the following properties that the ViewModel may interact with, and are easily bound within the XAML using a single binding.
+
+            - ItemSource (The items within the drop-down)
+            - SelectedItem (The currently selected item)
+            - SelectedValue \ SelectedValuePath ( The value of the SelectedItem, as an object, typically a property of the item )
+            - SelectedItemChanged event - Occurs and a new item is selected
+            - ItemSourceChanged event - Occurs when the itemsource is updated
+            - IsEnabled - Enable/Disable a control from within the ViewModel
+            - Visibilty / IsVisible - Toggle visibility of an item from within the ViewModel
+
+            ## XAML:
+            &lt;ComboBox MvvmControls:ControlDefinitions.ComboBoxDefinition="{Binding MyComboBox}" /&gt;
+        </Description>
+
+        <PackageReleaseNotes>
+            While this has targets for .net472, .net480, and .netCoreApp3.1, this library is set to use C#9.0 for Init-style properties, so your project may have to specify that `<LangVersion>9.0</LangVersion>` to use this.
+            .Net5.0 and beyond natively supports this library.
+        </PackageReleaseNotes>
+
+        <SignAssembly>true</SignAssembly>
+
+        <AssemblyOriginatorKeyFile>StrongName.pfx</AssemblyOriginatorKeyFile>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <None Include="..\LICENSE.txt" Link="LICENSE.txt">
+            <PackagePath></PackagePath>
+            <Pack>True</Pack>
+        </None>
+        <None Include="..\Readme.md" Link="Readme.md">
+            <PackagePath></PackagePath>
+            <Pack>True</Pack>
+        </None>
+    </ItemGroup>
+    
+    <ItemGroup>
+        <Compile Remove="WPF.Controls\OLD\IPV4Textbox.xaml.cs" />
+        <Compile Remove="WPF.Controls\OLD\TextBoxHelper.cs" />
+        <Compile Remove="WPF.Controls\OLD\TextboxValidation.xaml.cs" />
+        <Compile Remove="WPF.Controls\OLD\ValidationRuleProperty.cs" />
+    </ItemGroup>
+    <ItemGroup>
+        <Page Remove="WPF.Controls\OLD\IPV4Textbox.xaml" />
+    </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="CommunityToolkit.Mvvm" Version="7.0.1" />
+        <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.39" />
+        <PackageReference Include="System.Collections.Immutable" Version="6.0.0" />
+        <PackageReference Include="System.Drawing.Common" Version="6.0.0" />
+    </ItemGroup>
+    <ItemGroup>
+        <None Include="..\LICENSE.txt" Link="LICENSE.txt">
+            <PackagePath></PackagePath>
+            <Pack>True</Pack>
+        </None>
+        <None Include="WPF.Controls\OLD\IPV4Textbox.xaml" />
+        <None Include="WPF.Controls\OLD\IPV4Textbox.xaml.cs" />
+        <None Include="WPF.Controls\OLD\TextBoxHelper.cs" />
+        <None Include="WPF.Controls\OLD\TextboxValidation.xaml.cs" />
+        <None Include="WPF.Controls\OLD\ValidationRuleProperty.cs" />
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\Mvvm.ViewModel\Mvvm.IViewModel.csproj" />
+    </ItemGroup>
+    <ItemGroup>
+        <Page Update="WPF.Controls\IPV4TextBox\Generic.xaml">
+            <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
+            <SubType>Designer</SubType>
+        </Page>
+        <Page Update="WPF.Controls\NumericUpDowns\Generic.xaml">
+            <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
+            <SubType>Designer</SubType>
+        </Page>
+    </ItemGroup>
+
+</Project>
