+++ MvvmControlsTests/Mvvm/AsyncRelayCommandTests.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RFBCodeWorks.Mvvm;
+using RFBCodeWorks.Mvvm.Primitives;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -9,51 +10,94 @@ using System.Threading.Tasks;
 namespace RFBCodeWorks.Mvvm.Tests
 {
     [TestClass()]
-    public class AsyncRelayCommandTests2
+    public class AsyncRelayCommandTests : Primitives.Tests.CommandBaseTests
     {
-        private Task GetTask(int v) => GetTask(v, new CancellationTokenSource().Token);
-        private async Task GetTask(int loops, CancellationToken token)
+        private Task GetTask() => GetTask(new CancellationTokenSource().Token);
+        private async Task GetTask(CancellationToken token)
         {
             ErrorHandled = false;
             CancelHandled = false;
             int i = 0;
-            if (loops < i) throw new ArgumentException();
-            while (i < loops)
+            while (i < 25)
             {
+                if (ThrowError) throw new ArgumentException();
                 token.ThrowIfCancellationRequested();
-                await Task.Delay(1);
+                await Task.Delay(2);
                 i++;
             }
         }
 
-        private bool CanExecute(int value) => value > 0;
+        protected override CommandBase GetCommand() => new AsyncRelayCommand(cancelableExecute: GetTask);
+
+        private bool ThrowError;
+        private bool ExpectedCanExecute;
+
+        private bool CanExecute() => ExpectedCanExecute;
+
         private bool ErrorHandled;
         private bool CancelHandled;
-        private void HandleCancellation(int value) => CancelHandled = true;
+        private void HandleCancellation() => CancelHandled = true;
         private void HandleError(Exception e) => ErrorHandled = true;
+        internal static async Task AwaitCompletion(IAsyncRelayCommand cmd) { while (cmd.IsRunning) await Task.Delay(1); }
 
-        private async Task TestCanExecute(AsyncRelayCommand<int> cmd, bool methodSupplied = false)
+        private async Task TestCanExecute(AsyncRelayCommand cmd, bool methodSupplied = false)
         {
-            Assert.IsTrue(cmd.CanExecute(100), "\n\nCanExecute returned FALSE when expected TRUE");
-            if (methodSupplied) Assert.IsFalse(cmd.CanExecute(-100), "\n\nCanExecute returned TRUE with invalid input");
+            Console.WriteLine("Current Line# 41");
+            ExpectedCanExecute = true;
+            Assert.AreEqual(ExpectedCanExecute, cmd.CanExecute(), "\n\nCanExecute returned FALSE when expected TRUE");
+            if (methodSupplied)
+            {
+                ExpectedCanExecute = false;
+                Assert.IsFalse(cmd.CanExecute(), "\n\nCanExecute returned TRUE with invalid input");
+                ExpectedCanExecute = true;
+            }
 
             cmd.AllowConcurrentExecution = true;
-            _ = cmd.ExecuteAsync(10);
+            _ = cmd.ExecuteAsync();
             var task = cmd.ExecutionTask;
+            Console.WriteLine("Current Line# 54");
             Assert.IsNotNull(task, "ExecutionTask Property was not set!");
-            Assert.IsTrue(cmd.CanExecute(100), "\n\nCanExecute returned FALSE when expected TRUE");
+            Assert.IsTrue(cmd.CanExecute(), "\n\nCanExecute returned FALSE when expected TRUE");
             Assert.IsTrue(cmd.IsRunning, "\n\nIsRunning returned FALSE while task was running");
             await task;
             Assert.IsFalse(cmd.IsRunning, "\n\nIsRunning returned TRUE after task completed");
 
             cmd.AllowConcurrentExecution = false;
-            task = cmd.ExecuteAsync(10);
-            Assert.IsFalse(cmd.CanExecute(100), "\n\nCanExecute Returned TRUE while task was executing and AllowConcurrentExecution was false");
+            task = cmd.ExecuteAsync();
+            Assert.IsFalse(cmd.CanExecute(), "\n\nCanExecute Returned TRUE while task was executing and AllowConcurrentExecution was false");
             await task;
-            AsyncRelayCommandTests.TestIAsyncRelayCommand(cmd);
+
+            //Test Cancellation
+            Console.WriteLine("Current Line# 67");
+            cmd.AllowConcurrentExecution = true;
+            var t1 = cmd.ExecuteAsync();
+            var t2 = cmd.ExecuteAsync();
+            var t3 = cmd.ExecuteAsync();
+            Assert.AreNotSame(t1, t2, "\n\n Task1 and Task2 are the same take when expected to be different");
+            Assert.AreNotSame(t1, t3, "\n\n Task1 and Task3 are the same take when expected to be different");
+            Assert.AreNotSame(t2, t3, "\n\n Task2 and Task3 are the same take when expected to be different");
+            Assert.IsFalse(cmd.IsCancellationRequested, "\n\n IsCancellationRequested is TRUE prior to being cancelled");
+            if (cmd.CanBeCanceled)
+            {
+                Console.WriteLine("Current Line# 78");
+                cmd.SwallowCancellations = false;
+                cmd.Cancel();
+                Assert.IsTrue(cmd.IsCancellationRequested, "\n\nIsCancellationRequested is FALSE when expected to be true");
+                while (cmd.IsRunning) { await Task.Delay(3); }
+                await Task.Delay(10);
+                Assert.IsTrue(t3.IsCanceled, "\n\n Task3 was not cancelled");
+                Assert.IsTrue(t2.IsCanceled, "\n\n Task2 was not cancelled");
+                Assert.IsTrue(t1.IsCanceled, "\n\n Task1 was not cancelled");
+            }
+            else
+            {
+                Console.WriteLine("Current Line# 90");
+                await Task.WhenAll(t1, t2, t3);
+            }
+            TestIAsyncRelayCommand(cmd);
         }
 
-        private async Task TestExceptionThrowing(AsyncRelayCommand<int> cmd, bool hasErrorHandler, bool hasCancellationHandler, bool isCancellable = true)
+        private async Task TestExceptionThrowing(AsyncRelayCommand cmd, bool hasErrorHandler, bool hasCancellationHandler, bool isCancellable = true)
         {
             CancelHandled = false;
             ErrorHandled = false;
@@ -61,7 +105,9 @@ namespace RFBCodeWorks.Mvvm.Tests
             Task assertTask;
             if (!hasCancellationHandler && !hasErrorHandler && isCancellable)
             {
-                assertTask = Assert.ThrowsExceptionAsync<OperationCanceledException>(() => cmd.ExecuteAsync(100), "\n\nOperationCancelledExpection was not passed to caller!");
+                Console.WriteLine("Current Line# 103");
+                ThrowError = false;
+                assertTask = Assert.ThrowsExceptionAsync<OperationCanceledException>(() => cmd.ExecuteAsync(), "\n\nOperationCancelledExpection was not passed to caller!");
                 await Task.Delay(2);
                 cmd.Cancel();
                 await assertTask;
@@ -70,33 +116,39 @@ namespace RFBCodeWorks.Mvvm.Tests
             }
             else if (!hasErrorHandler)
             {
-                await Assert.ThrowsExceptionAsync<ArgumentException>(() => cmd.ExecuteAsync(-100), "\n\nArgumentException was not passed to caller!");
+                Console.WriteLine("Current Line# 113");
+                ThrowError = true;
+                await Assert.ThrowsExceptionAsync<ArgumentException>(() => cmd.ExecuteAsync(), "\n\nArgumentException was not passed to caller!");
                 Assert.IsFalse(CancelHandled, "\n\nCancel was handled unexpectedly");
                 Assert.IsFalse(ErrorHandled, "\n\nError was handled unexpectedly");
             }
             else if (!hasCancellationHandler && isCancellable)
             {
-                assertTask = cmd.ExecuteAsync(100);
+                Console.WriteLine("Current Line# 122");
+                ThrowError = false;
+                assertTask = cmd.ExecuteAsync();
                 await Task.Delay(2);
                 cmd.Cancel();
-                await AsyncRelayCommandTests.AwaitCompletion(cmd);
+                await AwaitCompletion(cmd);
                 Assert.IsFalse(CancelHandled, "\n\nCancel was handled unexpectedly");
                 Assert.IsFalse(ErrorHandled, "\n\nOperationCancelledExpection was not handled by the Error Handler!");
             }
+            ThrowError = false;
 
             //Test Cancellation Handling
             CancelHandled = false;
             if (hasCancellationHandler)
             {
+                Console.WriteLine("Current Line# 137");
                 Assert.IsFalse(cmd.CanBeCanceled, "\n\nCanBeCanceled returned TRUE while the task was not started");
-                Task v = cmd.ExecuteAsync(10);
+                Task v = cmd.ExecuteAsync();
                 Assert.IsTrue(cmd.CanBeCanceled, "\n\nCanBeCanceled returned FALSE while a Cancellable task was running");
                 await v;
                 Assert.IsFalse(CancelHandled, "\n\nCancellation was handled while task was not cancelled");
-                v = cmd.ExecuteAsync(100);
+                v = cmd.ExecuteAsync();
                 await Task.Delay(5);
                 cmd.Cancel();
-                await v;
+                await v.ContinueWith((t) => { });
                 Assert.IsTrue(CancelHandled, "\n\nCancellation was not handled when it was expected to be handled");
             }
 
@@ -104,44 +156,60 @@ namespace RFBCodeWorks.Mvvm.Tests
             ErrorHandled = false;
             if (hasErrorHandler)
             {
-                await cmd.ExecuteAsync(5);
+                Console.WriteLine("Current Line# 154");
+                ThrowError = false;
+                await cmd.ExecuteAsync();
                 Assert.IsFalse(ErrorHandled, "\n\nErrorHandled was raised unexpectedly");
-                _ = cmd.ExecuteAsync(-100);
+                ThrowError = true;
+                _ = cmd.ExecuteAsync();
                 Assert.IsTrue(ErrorHandled, "\n\nError was not handled when it was expected to be handled");
+                ThrowError = false;
             }
+        }
 
+        /// <summary>
+        /// Test the properties to ensure they don't throw a NotImplementedException
+        /// </summary>
+        /// <param name="cmd"></param>
+        internal static void TestIAsyncRelayCommand(IAsyncRelayCommand cmd)
+        {
+            var x = cmd.CanBeCanceled;
+            var y = cmd.ExecutionTask;
+            var z = cmd.IsCancellationRequested;
+            var a = cmd.IsRunning;
+            var b = cmd.RunningTasks;
         }
 
         [TestMethod("Cancelleable Task")]
         public void AsyncRelayCommandTest1()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(cancelableExecute: GetTask);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(cancelableExecute: GetTask);
             TestCanExecute(cmd).Wait();
-            TestExceptionThrowing(cmd, false, false).Wait();
+            TestExceptionThrowing(cmd, hasErrorHandler: false, hasCancellationHandler: false).Wait();
         }
 
         [TestMethod("Cancelleable Task + Error Handling")]
         public void AsyncRelayCommandTest2()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(cancelableExecute: GetTask, HandleError);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(cancelableExecute: GetTask, HandleError);
             TestCanExecute(cmd).Wait();
-            
+
             TestExceptionThrowing(cmd, hasErrorHandler: true, hasCancellationHandler: false).Wait();
         }
 
         [TestMethod("Cancelleable Task + Cancellation Handling")]
         public void AsyncRelayCommandTest3()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(cancelableExecute: GetTask, HandleCancellation);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(cancelableExecute: GetTask, HandleCancellation);
             TestCanExecute(cmd).Wait();
-            
+
             TestExceptionThrowing(cmd, hasErrorHandler: false, hasCancellationHandler: true).Wait();
         }
 
         [TestMethod("Cancelleable Task + CanExecuteValidation")]
         public void AsyncRelayCommandTest4()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(cancelableExecute: GetTask, CanExecute);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(cancelableExecute: GetTask, CanExecute);
             TestCanExecute(cmd, true).Wait();
             TestExceptionThrowing(cmd, hasErrorHandler: false, hasCancellationHandler: false).Wait();
         }
@@ -149,35 +217,35 @@ namespace RFBCodeWorks.Mvvm.Tests
         [TestMethod("Cancelleable Task + CanExecuteValidation + Error Handling")]
         public void AsyncRelayCommandTest5()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(cancelableExecute: GetTask, CanExecute, HandleError);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(cancelableExecute: GetTask, CanExecute, HandleError);
             TestCanExecute(cmd, true).Wait();
-            
+
             TestExceptionThrowing(cmd, hasErrorHandler: true, hasCancellationHandler: false).Wait();
         }
 
         [TestMethod("Cancelleable Task + CanExecuteValidation + Cancellation Handling")]
         public void AsyncRelayCommandTest6()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(cancelableExecute: GetTask, CanExecute, HandleCancellation);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(cancelableExecute: GetTask, CanExecute, HandleCancellation);
             TestCanExecute(cmd, true).Wait();
-            
+
             TestExceptionThrowing(cmd, hasErrorHandler: false, hasCancellationHandler: true).Wait();
         }
 
         [TestMethod("Cancelleable Task + CanExecuteValidation + Error Handling + Cancellation Handling")]
         public void AsyncRelayCommandTest7()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(cancelableExecute: GetTask, CanExecute, HandleError, HandleCancellation);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(cancelableExecute: GetTask, CanExecute, HandleError, HandleCancellation);
             TestCanExecute(cmd, true).Wait();
-            
-            
+
+
             TestExceptionThrowing(cmd, hasErrorHandler: true, hasCancellationHandler: true).Wait();
         }
 
         [TestMethod("Non-Cancellable Task")]
         public void AsyncRelayCommandTest11()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(execute: GetTask);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(execute: GetTask);
             TestCanExecute(cmd, false).Wait();
             TestExceptionThrowing(cmd, hasErrorHandler: false, hasCancellationHandler: false, isCancellable: false).Wait();
         }
@@ -185,7 +253,7 @@ namespace RFBCodeWorks.Mvvm.Tests
         [TestMethod("Non-Cancellable Task + CanExecuteValidation")]
         public void AsyncRelayCommandTest9()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(execute: GetTask, CanExecute);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(execute: GetTask, CanExecute);
             TestCanExecute(cmd, true).Wait();
             TestExceptionThrowing(cmd, hasErrorHandler: false, hasCancellationHandler: false, isCancellable: false).Wait();
         }
@@ -193,18 +261,18 @@ namespace RFBCodeWorks.Mvvm.Tests
         [TestMethod("Non-Cancellable Task + Error Handling")]
         public void AsyncRelayCommandTest8()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(execute: GetTask, HandleError);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(execute: GetTask, HandleError);
             TestCanExecute(cmd, false).Wait();
-            
+
             TestExceptionThrowing(cmd, hasErrorHandler: true, hasCancellationHandler: false, isCancellable: false).Wait();
         }
 
         [TestMethod("Non-Cancellable Task + CanExecuteValidation + Error Handling")]
         public void AsyncRelayCommandTest10()
         {
-            AsyncRelayCommand<int> cmd = new AsyncRelayCommand<int>(execute: GetTask, CanExecute, HandleError);
+            AsyncRelayCommand cmd = new AsyncRelayCommand(execute: GetTask, CanExecute, HandleError);
             TestCanExecute(cmd, true).Wait();
-            
+
             TestExceptionThrowing(cmd, hasErrorHandler: true, hasCancellationHandler: false, isCancellable: false).Wait();
         }
     }
