+++ MvvmControlsTests/Mvvm/CheckBoxDefinitionTests.cs
