+++ MvvmControlsTests/Mvvm/ObjectViewModelTests.cs
