+++ MvvmControlsTests/Mvvm/Primitives/ToggleButtonDefinitionTests.cs
