+++ MvvmControlsTests/Mvvm/RadioButtonDefinitionTests.cs
