+++ MvvmControlsTests/MvvmControlsTests.csproj
@@ -19,7 +19,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MvvmControls\MvvmControls.csproj" />
+    <ProjectReference Include="..\MvvmControls\Mvvm.csproj" />
   </ItemGroup>
 
 </Project>
