+++ MvvmControlsTests/XmlLinq/Controls/XComboBoxElementWrapperTests.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RFBCodeWorks.Mvvm.Tests;
+using RFBCodeWorks.Mvvm.Tests.Helpers;
 using System.Windows.Controls;
 
 namespace RFBCodeWorks.Mvvm.XmlLinq.Controls.Tests
